+++ server/libs/modules/components/ai/vectorstore/oracle/src/test/resources/definition/oracle_v1.json
@@ -18,7 +18,6 @@
     "processErrorResponse": null,
     "properties": [ ],
     "resumePerform": null,
-    "suspendPerform": null,
     "title": "Load Data",
     "workflowNodeDescription": null
   }, {
@@ -62,7 +61,6 @@
       "type": "STRING"
     } ],
     "resumePerform": null,
-    "suspendPerform": null,
     "title": "Search Data",
     "workflowNodeDescription": null
   } ],
